change components and move context to prevent cycle dependencies
