+++ src/App.js
@@ -1,53 +1,26 @@
-import React, { Suspense, lazy, createContext } from 'react';
+import React, { Suspense, lazy } from 'react';
 import { BrowserRouter, Route, Routes, Navigate } from 'react-router-dom';
 import ReactGA from 'react-ga';
 import 'react-calendar/dist/Calendar.css';
 import { Modal, ModalContent, ModalOverlay } from '@chakra-ui/core';
 import { gaTrackingId } from './config';
 import Loader from './components/ui/loader';
+import { mainReducer, mainState, MainContext } from './context/main-context';
 
 const Home = lazy(() => import('./pages/home'));
 const Campaign = lazy(() => import('./pages/campaign'));
 const Redirecting = lazy(() => import('./pages/redirecting'));
 const Admin = lazy(() => import('./pages/admin'));
 
-export const WalletContext = createContext();
-
-const initialState = {
-  wallet: '',
-  modal: { isOpen: false, overlay: true, closable: true, otherProps: {} },
-};
-
-const reducer = (state, action) => {
-  switch (action.type) {
-    case 'SET_WALLET':
-      localStorage.setItem('wallet', JSON.stringify(action.payload));
-      return {
-        ...state,
-        wallet: action.payload,
-      };
-    case 'SET_MODAL':
-      return {
-        ...state,
-        modal: {
-          ...initialState.modal,
-          ...action.payload,
-        },
-      };
-    default:
-      return state;
-  }
-};
-
 ReactGA.initialize(gaTrackingId);
 
 function App() {
-  const [state, dispatch] = React.useReducer(reducer, initialState);
+  const [state, dispatch] = React.useReducer(mainReducer, mainState);
   React.useEffect(() => {
     ReactGA.pageview(window.location.pathname + window.location.search);
   }, []);
   return (
-    <WalletContext.Provider value={{ state, dispatch }}>
+    <MainContext.Provider value={{ state, dispatch }}>
       <BrowserRouter>
         <Suspense fallback={<Loader isFull />}>
           <Routes>
@@ -83,7 +56,7 @@ function App() {
           <ModalContent>{state.modal.content}</ModalContent>
         </Modal>
       )}
-    </WalletContext.Provider>
+    </MainContext.Provider>
   );
 }
